requirements.txt
imageio==2.37.0
lazy_loader==0.4
networkx==3.5
numpy==2.3.2
packaging==25.0
pillow==11.3.0
scikit-image==0.25.2
scipy==1.16.1
tifffile==2025.6.11
tqdm==4.67.1
